refactor: move product domain modules into polylogue/products/
Move 8 product modules from package root into polylogue/products/:
- archive_product_models → products/archive_models
- archive_product_rollups → products/archive_rollups
- archive_product_summaries → products/archive_summaries
- archive_products → products/archive
- archive_resume → products/resume
- authored_payloads → products/authored_payloads
- product_export_bundles → products/export_bundles
- product_readiness → products/readiness

Remove cli/products_workflow.py compatibility shim.
Update topology-target.yaml (#414 cells → stable).

Ref #414
