+++ polylogue/api/products.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Protocol
 
-from polylogue.archive_products import (
+from polylogue.products.archive import (
     ArchiveDebtProduct,
     ArchiveDebtProductQuery,
     CostRollupProduct,
