+++ polylogue/scenarios/projections.py
@@ -6,7 +6,7 @@ from collections.abc import Iterable
 from dataclasses import asdict, dataclass, field
 from enum import Enum
 
-from polylogue.authored_payloads import PayloadDict, PayloadMap
+from polylogue.products.authored_payloads import PayloadDict, PayloadMap
 
 from .metadata import ScenarioMetadata
