+++ polylogue/showcase/catalog_loader.py
@@ -7,7 +7,7 @@ from collections.abc import Callable
 from dataclasses import dataclass
 from pathlib import Path
 
-from polylogue.authored_payloads import (
+from polylogue.products.authored_payloads import (
     PayloadDict,
     payload_bool,
     payload_float,
