+++ polylogue/showcase/showcase_report_payloads.py
@@ -6,7 +6,7 @@ import json
 from datetime import datetime, timezone
 from pathlib import Path
 
-from polylogue.authored_payloads import PayloadDict
+from polylogue.products.authored_payloads import PayloadDict
 from polylogue.showcase.report_models import (
     ShowcaseExerciseRecord,
     ShowcaseSessionRecord,
