+++ polylogue/storage/products/session/timeline_rows.py
@@ -4,17 +4,17 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from polylogue.archive_product_models import (
-    SessionPhaseEvidencePayload,
-    SessionPhaseInferencePayload,
-    WorkEventEvidencePayload,
-    WorkEventInferencePayload,
-)
 from polylogue.lib.conversation.extraction import WorkEvent
 from polylogue.lib.hashing import hash_text
 from polylogue.lib.phase.extraction import SessionPhase
 from polylogue.lib.session.documents import WorkEventDocument
 from polylogue.lib.session.session_profile import SessionProfile
+from polylogue.products.archive_models import (
+    SessionPhaseEvidencePayload,
+    SessionPhaseInferencePayload,
+    WorkEventEvidencePayload,
+    WorkEventInferencePayload,
+)
 from polylogue.storage.products.session.profiles import (
     event_fallback,
     event_summary,
