+++ tests/unit/core/test_facade_api.py
@@ -8,7 +8,8 @@ import pytest
 
 from polylogue import Polylogue
 from polylogue.api import ArchiveStats
-from polylogue.archive_products import (
+from polylogue.lib.semantic.content_projection import ContentProjectionSpec
+from polylogue.products.archive import (
     ArchiveDebtProductQuery,
     CostRollupProductQuery,
     DaySessionSummaryProductQuery,
@@ -20,7 +21,6 @@ from polylogue.archive_products import (
     WeekSessionSummaryProductQuery,
     WorkThreadProductQuery,
 )
-from polylogue.lib.semantic.content_projection import ContentProjectionSpec
 from tests.infra.builders import make_conv, make_msg
 from tests.infra.storage_records import ConversationBuilder, make_conversation, make_message
