+++ tests/unit/core/test_repo_identity.py
@@ -7,8 +7,6 @@ from pathlib import Path
 
 import pytest
 
-from polylogue.archive_product_models import DaySessionSummaryPayload
-from polylogue.archive_product_summaries import aggregate_day_session_summary_products
 from polylogue.lib.action_event.action_events import ActionEvent
 from polylogue.lib.conversation.attribution import extract_attribution, extract_attribution_from_action_events
 from polylogue.lib.message.messages import MessageCollection
@@ -23,6 +21,8 @@ from polylogue.lib.roles import Role
 from polylogue.lib.session.session_profile import SessionProfile, build_session_profile
 from polylogue.lib.session.session_summaries import summarize_day
 from polylogue.lib.viewport.viewports import ToolCategory
+from polylogue.products.archive_models import DaySessionSummaryPayload
+from polylogue.products.archive_summaries import aggregate_day_session_summary_products
 from polylogue.storage.runtime import DaySessionSummaryRecord
 from polylogue.types import ConversationId, Provider
