+++ tests/unit/showcase/test_report.py
@@ -15,8 +15,8 @@ from pathlib import Path
 from hypothesis import given, settings
 from hypothesis import strategies as st
 
-from polylogue.authored_payloads import require_payload_mapping
 from polylogue.lib.outcomes import OutcomeCheck, OutcomeStatus
+from polylogue.products.authored_payloads import require_payload_mapping
 from polylogue.scenarios import CorpusSpec, polylogue_execution
 from polylogue.schemas.audit_models import AuditReport
 from polylogue.schemas.verification_models import ArtifactProofReport, ProviderArtifactProof
